[Refactor] 런타임 성능, 컴파일 타임 성능 최적화 및 폴더 생성
동적 디스패치 보다는 정적 디스패치를 지향하기 위해, 상속/참조/identity에 따른 제어/Objective-C와의 상호운용이 필요 없는 class는 final class로 변경한다.
폴더를 생성해 연관 소스파일끼리 관리한다.

Fixes: https://camp.yagom-academy.kr/camp/65cc2b983173745319e0464c/projects/65cc320e3173745319e04663
